Refactor: rename createAndUpdateDailyRecords to manageDailyRecords to resolve Heroku deploy issue
